Rename query provider file
